server/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-groq
langchain-community
langchain-huggingface
python-dotenv
faiss-cpu
beautifulsoup4
requests
sentence-transformers
pydantic